[generator-emakinacee-react] align templates with single setup
