--- packages/generator-emakinacee-react/generators/app/templates/base/static/public/index.html
@@ -1,34 +1,0 @@
-<!doctype html>
-<html lang="en">
-    <head>
-        <meta charset="utf-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-        <link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico">
-        <!--
-            Notice the use of %PUBLIC_URL% in the tag above.
-            It will be replaced with the URL of the `public` folder during the build.
-            Only files inside the `public` folder can be referenced from the HTML.
-
-            Unlike "/favicon.ico" or "favicon.ico", "%PUBLIC_URL%/favicon.ico" will
-            work correctly both with client-side routing and a non-root public URL.
-            Learn how to configure a non-root public URL by running `npm run build`.
-        -->
-        <title>React App</title>
-    </head>
-    <body>
-        <noscript>
-            You need to enable JavaScript to run this app.
-        </noscript>
-        <div id="root"></div>
-        <!--
-            This HTML file is a template.
-            If you open it directly in the browser, you will see an empty page.
-
-            You can add webfonts, meta tags, or analytics to this file.
-            The build step will place the bundled scripts into the <body> tag.
-
-            To begin the development, run `npm start`.
-            To create a production bundle, use `npm run build`.
-        -->
-    </body>
-</html>
