--- packages/generator-emakinacee-react/generators/app/templates/base/static/src/index.js
@@ -1,6 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import App from './components/App/App';
-import './index.css';
-
-ReactDOM.render(<App />, document.getElementById('root'));
