+++ packages/generator-emakinacee-react/generators/app/templates/dot/.eslintrc
