+++ packages/generator-emakinacee-react/generators/app/templates/dynamic/manifest.json
