+++ packages/generator-emakinacee-react/generators/app/templates/dynamic/package.json
