--- packages/generator-emakinacee-react/generators/app/templates/shared/static/src/shared/services/ApiService.js
@@ -1,31 +1,0 @@
-/**
- * API SERVICE
- * This is just an example service to show where services
- * live in the folder structure.
- *
- * Feel free to extend or remove it.
- */
-class ApiService {
-
-    constructor() {
-        this.foo = {};
-    }
-
-    get() {
-        return this.foo;
-    }
-
-    put() {
-        return this.foo;
-    }
-
-    post() {
-        return this.foo;
-    }
-
-    delete() {
-        return this.foo;
-    }
-}
-
-export default ApiService;
