--- packages/generator-emakinacee-react/generators/app/templates/shared/static/src/shared/services/ApiService.spec.js
@@ -1,5 +1,0 @@
-import ApiService from './ApiService';
-
-it('BlubService is defined', () => {
-    expect(new ApiService()).toBeInstanceOf(ApiService);
-});
