--- packages/generator-emakinacee-react/generators/app/templates/spa/dynamic/package.json
@@ -1,13 +1,0 @@
-{
-  "name": "<%= projectName %>",
-  "version": "0.0.0",
-  "private": true,
-  "description": "",
-  "scripts": {
-    "start": "react-scripts-emakinacee start",
-    "build": "react-scripts-emakinacee build",
-    "test": "react-scripts-emakinacee test --env=jsdom",
-    "serve:build": "serve ./build",
-    "eject": "react-scripts-emakinacee eject"
-  }
-}
