--- packages/generator-emakinacee-react/generators/app/templates/spa/static/src/components/App/App.scss
@@ -1,28 +1,0 @@
-@import "../../scss/base-imports";
-
-.c-app {
-    @include mq($from: tablet) {
-        text-align: center;
-    }
-
-    &__logo {
-        animation: app-logo-spin infinite 20s linear;
-        height: rem(80);
-    }
-
-    &__header {
-        background-color: map_get($color-greyscale, darkest);
-        height: rem(150);
-        padding: rem(20);
-        color: map_get($color-greyscale, white);
-    }
-
-    &__intro {
-        font-size: large;
-    }
-}
-
-@keyframes app-logo-spin {
-    from { transform: rotate(0deg); }
-    to { transform: rotate(360deg); }
-}
