--- packages/generator-emakinacee-react/generators/app/templates/spa/static/src/components/App/App.spec.js
@@ -1,16 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import {Container} from 'cerebral/react';
-import controller from '../../controller';
-import App from './App';
-
-it('renders without crashing', () => {
-    const div = document.createElement('div');
-
-    ReactDOM.render(
-        <Container controller={controller} >
-            <App />
-        </Container>,
-        div
-    );
-});
