--- packages/generator-emakinacee-react/generators/app/templates/spa/static/src/components/Minions/Minions.scss
@@ -1,14 +1,0 @@
-@import "../../scss/base-imports";
-
-.c-minions {
-
-    ul {
-        padding: 0;
-        margin: rem(30 0 0);
-        list-style: none;
-    }
-
-    li {
-        padding: rem(10 0);
-    }
-}
