--- packages/generator-emakinacee-react/generators/app/templates/spa/static/src/modules/App/AppModule.js
@@ -1,16 +1,0 @@
-import {state} from 'cerebral/tags';
-import {set} from 'cerebral/operators';
-
-export default {
-    state: {
-        $loading: true
-    },
-    signals: {
-        bootstrap: [
-            set(state`app.$loading`, false)
-        ],
-        routed: [
-            set(state`app.$selectedView`, 'Home')
-        ]
-    }
-};
