--- packages/generator-emakinacee-react/generators/app/templates/spa/static/src/modules/Minions/MinionsModule.js
@@ -1,22 +1,0 @@
-import {set, push} from 'cerebral/operators';
-import {state, props} from 'cerebral/tags';
-import fetchMinion from './actions/fetchMinionAction';
-
-export default {
-    state: {
-        available: [
-            {id: 1, name: 'Bob'},
-            {id: 2, name: 'Stuart'},
-            {id: 3, name: 'Kevin'}
-        ]
-    },
-    signals: {
-        addMinionClicked: [
-            fetchMinion,
-            push(state`minions.available`, props`apiMinion`)
-        ],
-        routed: [
-            set(state`app.$selectedView`, 'Minions')
-        ]
-    }
-};
