--- packages/generator-emakinacee-react/generators/app/templates/spa/static/src/modules/Minions/MinionsModule.spec.js
@@ -1,6 +1,0 @@
-import MinionsModule from './MinionsModule';
-
-it('Modules exports object with state and signals', () => {
-    expect(MinionsModule).toHaveProperty('state');
-    expect(MinionsModule).toHaveProperty('signals');
-});
