+++ packages/generator-emakinacee-react/generators/app/templates/static/public/favicon.ico
