+++ packages/generator-emakinacee-react/generators/app/templates/static/public/index.html
