+++ packages/generator-emakinacee-react/generators/app/templates/static/src/components/App/App.scss
