+++ packages/generator-emakinacee-react/generators/app/templates/static/src/components/Home/Home.js
@@ -1,10 +1,12 @@
 import React from 'react';
+import { connect } from '@cerebral/react';
+import { state } from 'cerebral/tags';
 import './Home.scss';
 
-const Home = () => {
+const Home = ({ title }) => {
     return (
         <div className="c-home">
-            <a href="/minions">Watch our minions</a>
+            <h1>{title}</h1>
             <p className="c-home__intro">
                 To get started, edit <code>src/App.js</code> and save to reload.
             </p>
@@ -12,4 +14,9 @@ const Home = () => {
     );
 };
 
-export default Home;
+export default connect(
+    {
+        title: state`app.title`,
+    },
+    Home
+);
