+++ packages/generator-emakinacee-react/generators/app/templates/static/src/components/Home/Home.scss
