+++ packages/generator-emakinacee-react/generators/app/templates/static/src/index.scss
