+++ packages/generator-emakinacee-react/generators/app/templates/static/src/modules/App/AppModule.js
@@ -1,0 +1,7 @@
+export default {
+    state: {
+        title: 'Hello World'
+    },
+    signals: {
+    }
+};
