+++ packages/generator-emakinacee-react/generators/app/templates/static/src/modules/App/AppModule.spec.js
