+++ packages/generator-emakinacee-react/generators/app/templates/static/src/registerServiceWorker.js
