+++ packages/generator-emakinacee-react/generators/app/templates/static/src/scss/_base-imports.scss
