+++ packages/generator-emakinacee-react/generators/app/templates/static/src/scss/settings/_settings.typography.scss
